agora testando tudo de novo.

+++ test_libft/libft/ft_strlcat.c
@@ -6,7 +6,7 @@
 /*   By: lsarraci <lsarraci@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/07/14 12:20:39 by lsarraci          #+#    #+#             */
-/*   Updated: 2025/07/22 14:48:15 by lsarraci         ###   ########.fr       */
+/*   Updated: 2025/07/23 12:37:20 by lsarraci         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,8 +22,8 @@ size_t	ft_strlcat( char *dst, const char *src, size_t size)
 
 	if (!dst)
 		return (0);
-	dst_len = (size_t)ft_strlen(dst);
-	src_len = (size_t)ft_strlen((char *)src);
+	dst_len = ft_strlen(dst);
+	src_len = ft_strlen((char *)src);
 	j = 0;
 	i = dst_len;
 	if (!src)
